.history/requirements_20250419173845.txt
pandas>=1.3.4
openpyxl>=3.0.9
pillow>=8.3.1
PyQt5>=5.15.7
pytest>=6.2.5
pyinstaller>=4.5.1
typing-extensions>=3.10.0
colorlog>=6.0.0
pyyaml>=6.0
psutil>=5.9.0
pystray>=0.19.3
colorama>=0.4.4
pywin32>=303 